Fix user registration
