--- src/main/java/org/example/advs/controller/ControllerUtils.java
@@ -1,16 +1,0 @@
-package org.example.advs.controller;
-
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
-
-import java.util.Map;
-import java.util.stream.Collector;
-import java.util.stream.Collectors;
-
-public class ControllerUtils {
-    static Map<String, String> getErrors(BindingResult bindingResult){
-        Collector<FieldError, ?, Map<String, String>> collector = Collectors.toMap(fieldError ->
-                fieldError.getField() + "Error", FieldError::getDefaultMessage);
-        return bindingResult.getFieldErrors().stream().collect(collector);
-    }
-}
