--- src/main/java/org/example/advs/domain/Role.java
@@ -1,13 +1,0 @@
-package org.example.advs.domain;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum Role implements GrantedAuthority {
-    USER, ADMIN;
-
-    @Override
-    public String getAuthority() {
-        return name();
-    }
-
-}
