--- src/main/java/org/example/advs/repo/CatalogRepo.java
@@ -1,8 +1,0 @@
-package org.example.advs.repo;
-
-import org.example.advs.domain.Catalog;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CatalogRepo extends JpaRepository<Catalog, Long> {
-    Catalog findByTitle(String title);
-}
