+++ src/main/java/org/fetisman/ads/Application.java
@@ -1,0 +1,13 @@
+package org.fetisman.ads;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}
