+++ src/main/java/org/fetisman/ads/domain/CaptchaResponseDto.java
@@ -1,0 +1,30 @@
+package org.fetisman.ads.domain;
+
+import com.fasterxml.jackson.annotation.JsonAlias;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.Set;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class CaptchaResponseDto {
+    private boolean success;
+
+    @JsonAlias("error-codes")
+    private Set<String> errorCodes;
+
+    public boolean isSuccess() {
+        return success;
+    }
+
+    public void setSuccess(boolean success) {
+        this.success = success;
+    }
+
+    public Set<String> getErrorCodes() {
+        return errorCodes;
+    }
+
+    public void setErrorCodes(Set<String> errorCodes) {
+        this.errorCodes = errorCodes;
+    }
+}
