+++ src/main/java/org/fetisman/ads/domain/Role.java
@@ -1,0 +1,13 @@
+package org.fetisman.ads.domain;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum Role implements GrantedAuthority {
+    USER, ADMIN;
+
+    @Override
+    public String getAuthority() {
+        return name();
+    }
+
+}
