+++ src/main/java/org/fetisman/ads/repo/UserRepo.java
@@ -1,0 +1,10 @@
+package org.fetisman.ads.repo;
+
+import org.fetisman.ads.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepo extends JpaRepository<User, Long> {
+    User findByUsername(String userName);
+
+    User findByActivationCode(String code);
+}
